refactor: Chartコンポーネントを親要素として、複数チャートを表示できるようにした
