+++ frontend/src/features/Chart/Chart.tsx
@@ -1,4 +1,3 @@
-import { Bar } from "react-chartjs-2";
 import {
   Chart as ChartJS,
   CategoryScale,
@@ -11,6 +10,7 @@ import {
 import { GetSprintList } from "../sprintlist/GetConstSprintList";
 import { useEffect, useState } from "react";
 import { FetchPullRequests } from "../pullrequestlist/PullRequestsFetcher";
+import { MetricsChart } from "./MetricsChart";
 
 ChartJS.register(
   CategoryScale,
@@ -21,68 +21,18 @@ ChartJS.register(
   Legend
 );
 
-const chartOptions = {
-  plugins: {
-    title: {
-      display: true,
-      text: 'スプリントあたりの開発メトリクスチャート',
-    },
-    annotation: { 
-      annotations: [{ 
-        type: 'line', // 線を描画
-        id: '2dayBorder', 
-        mode: 'horizontal', // 線を水平に引く
-        scaleID: 'y-axis-0', 
-        value: 86400, // 基準となる数値
-        borderWidth: 10, // 基準線の太さ
-        borderColor: 'red'  // 基準線の色
-      }] 
-    },  
-  },
-  responsive: true,
-  scales: {
-    x: {
-      stacked: true,
-    },
-    y: {
-      stacked: true,
-    },
-  },
-}
 
-type Metrics = {
+export type Metrics = {
   sprintId: number,
   score: number,
 }
 
 export const Chart = () => {
   const sprintList = GetSprintList()
-  const labels = sprintList.map((sprint) => sprint.id)
 
   const [untilFirstReviewedList, setUntilFirstReviewedList] = useState<Metrics[]>([]);
   const [untilLastApprovedList, setUntilLastApprovedList] = useState<Metrics[]>([]);
   const [untilMergedList, setUntilMergedList] = useState<Metrics[]>([]);
-
-  const datasets = {
-    labels,
-    datasets: [
-      {
-        label: 'レビューまでにかかった時間',
-        data: untilFirstReviewedList.map((metrics) => metrics.score),
-        backgroundColor: 'rgb(255, 99, 132)',
-      },
-      {
-        label: '最後のapproveまでにかかった時間',
-        data: untilLastApprovedList.map((metrics) => metrics.score),
-        backgroundColor: 'rgb(75, 192, 192)',
-      },
-      {
-        label: 'マージまでにかかった時間',
-        data: untilMergedList.map((metrics) => metrics.score),
-        backgroundColor: 'rgb(53, 162, 235)',
-      },
-    ],
-  };
 
   useEffect(() => {
     // スプリント毎のPRをチャートに反映する
@@ -119,10 +69,7 @@ export const Chart = () => {
 
 return (
     <>
-      {datasets
-       ? <Bar options={chartOptions} data={datasets} />
-       : <div>Loading...</div>
-      }
+      <MetricsChart sprintList={sprintList} untilFirstReviewedList={untilFirstReviewedList} untilLastApprovedList={untilLastApprovedList} untilMergedList={untilMergedList} />
     </>
   )
 };
